chore: refactor reaction role
