+++ src/commands/reaction-role/remove/slash.ts
@@ -1,13 +1,9 @@
-import { RoleReactionEvent, SlashCommand } from "types/common"
+import { SlashCommand } from "types/common"
 import { SLASH_PREFIX as PREFIX } from "utils/constants"
 import { composeEmbedMessage2, getErrorEmbed } from "ui/discord/embed"
 import { CommandInteraction } from "discord.js"
-import config from "adapters/config"
-import ChannelLogger from "logger/channel"
-import { BotBaseError, InternalError } from "errors"
-import { logger } from "logger"
 import { SlashCommandSubcommandBuilder } from "@discordjs/builders"
-import { isDiscordMessageLink } from "utils/common"
+import { handleRoleRemove } from "./processor"
 
 const command: SlashCommand = {
   name: "remove",
@@ -30,7 +26,7 @@ const command: SlashCommand = {
           .setDescription("emoji which you want to configure for role")
           .setRequired(false)
       )
-      .addStringOption((option) =>
+      .addRoleOption((option) =>
         option
           .setName("role")
           .setDescription("role which you want to configure")
@@ -50,155 +46,14 @@ const command: SlashCommand = {
         },
       }
     }
-    let description = ""
-
-    // Validate command syntax
-    const role = interaction.options.getString("role", false)
-    const emoji = interaction.options.getString("emoji", false)
-    if ((role == null) != (emoji == null)) {
-      return {
-        messageOptions: {
-          embeds: [
-            composeEmbedMessage2(interaction, {
-              usage: `To remove a specific configuration in a message\n${PREFIX}reactionrole remove <message_id> <emoji> <role>\n\nTo clear all configurations in a message\n${PREFIX}reactionrole remove <message_id>`,
-              examples: `${PREFIX}reactionrole remove 967107573591457832 ✅ @Visitor\n${PREFIX}reactionrole remove 967107573591457832`,
-            }),
-          ],
-        },
-      }
-    }
-
-    // Validate message_id
     const messageLink = interaction.options.getString("message_link", true)
-    if (!isDiscordMessageLink(messageLink)) {
-      throw new InternalError({
-        message: interaction,
-        description:
-          "Invalid message link, use `$help rr` to learn how to get message link",
-      })
-    }
-
-    const [guildId, channelId, messageId] = messageLink.split("/").slice(-3)
-    if (guildId !== interaction.guildId) {
-      throw new InternalError({
-        message: interaction,
-        description:
-          "Guild ID invalid, please choose a message belongs to your guild",
-      })
-    }
-
-    const channel = interaction.guild.channels.cache.get(channelId) // user already has message in the channel => channel in cache
-    if (!channel || !channel.isText()) {
-      throw new InternalError({
-        message: interaction,
-        description: "Channel not found",
-      })
-    }
-
-    const message = await channel.messages.fetch(messageId).catch(() => null)
-    if (!message) {
-      throw new InternalError({
-        message: interaction,
-        description: "Message not found",
-      })
-    }
-
-    if (!message || !messageId) {
-      return {
-        messageOptions: {
-          embeds: [
-            getErrorEmbed({
-              description: "Message not found",
-              originalMsgAuthor: interaction.user,
-            }),
-          ],
-        },
-      }
-    }
-
-    let requestData: RoleReactionEvent | null = null
-    if (!emoji && !role) {
-      // Remove all reaction from configured message
-      requestData = {
-        guild_id: interaction.guild.id,
-        message_id: message.id,
-        reaction: "",
-        role_id: "",
-        channel_id: channel.id,
-      }
-    } else if (emoji && role) {
-      // Remove a specific reaction
-      const roleId = role.replace(/\D/g, "")
-      const guildRole = await interaction.guild.roles.fetch(roleId)
-      if (!guildRole || !roleId) {
-        return {
-          messageOptions: {
-            embeds: [
-              getErrorEmbed({
-                description: "Role not found",
-                originalMsgAuthor: interaction.user,
-              }),
-            ],
-          },
-        }
-      }
-      requestData = {
-        guild_id: interaction.guild.id,
-        message_id: message.id,
-        reaction: emoji,
-        role_id: roleId,
-        channel_id: channel.id,
-      }
-    }
-
-    if (requestData) {
-      try {
-        const res = await config.removeReactionConfig(requestData)
-        if (res.ok) {
-          const { reaction, role_id } = requestData
-          if (reaction && role_id) {
-            description = `Reaction ${reaction} is removed for <@&${role_id}>.`
-
-            const emojiSplit = reaction.split(":")
-            const reactionEmoji =
-              emojiSplit.length === 1 ? reaction : reaction.replace(/\D/g, "")
-            message.reactions.cache
-              .get(reactionEmoji)
-              ?.remove()
-              .catch((e) => {
-                logger.info(e)
-              })
-          } else {
-            description = `All reaction role configurations for this message is now clear.`
-            message.reactions.removeAll().catch((e) => {
-              logger.info(e)
-            })
-          }
-        } else {
-          throw new InternalError({
-            message,
-            title: "Unsuccessful",
-            description: `You haven't set this reaction role yet. To set a new one, run \`\`\`${PREFIX}rr set <message_link> <emoji> <role>\`\`\`\n You can remove it later using \`${PREFIX}rr remove\`.`,
-          })
-        }
-      } catch (error) {
-        ChannelLogger.log(error as BotBaseError)
-        throw new InternalError({
-          message,
-          title: "Unsuccessful",
-          description: `You haven't set this reaction role yet. To set a new one, run \`\`\`${PREFIX}rr set <message_link> <emoji> <role>\`\`\`\n You can remove it later using \`${PREFIX}rr remove\`.`,
-        })
-      }
-    }
+    const role = interaction.options.getRole("role", false)
+    const emoji = interaction.options.getString("emoji", false)
+    const args = ["", "", messageLink, emoji ?? "", role?.id ?? ""]
 
     return {
       messageOptions: {
-        embeds: [
-          composeEmbedMessage2(interaction, {
-            author: ["Reaction roles", interaction.guild.iconURL()],
-            description,
-          }),
-        ],
+        ...(await handleRoleRemove(args, interaction)),
       },
     }
   },
